+++ src/CSharp.UnionTypes.CLI/AssemblyInfo.fs
@@ -3,7 +3,7 @@ namespace System
 open System.Reflection
 
 [<assembly: AssemblyTitleAttribute("CSharp.UnionTypes.CLI")>]
-[<assembly: AssemblyProductAttribute("BrightSword.CSharpExtensions.DiscriminatedUnion")>]
+[<assembly: AssemblyProductAttribute("CSharp.UnionTypes")>]
 [<assembly: AssemblyDescriptionAttribute("DSL, Parser & Code Generator for Discriminated Union Types (Sum Types) in C#")>]
 [<assembly: AssemblyVersionAttribute("1.0.0")>]
 [<assembly: AssemblyFileVersionAttribute("1.0.0")>]
@@ -11,7 +11,7 @@ do ()
 
 module internal AssemblyVersionInformation =
     let [<Literal>] AssemblyTitle = "CSharp.UnionTypes.CLI"
-    let [<Literal>] AssemblyProduct = "BrightSword.CSharpExtensions.DiscriminatedUnion"
+    let [<Literal>] AssemblyProduct = "CSharp.UnionTypes"
     let [<Literal>] AssemblyDescription = "DSL, Parser & Code Generator for Discriminated Union Types (Sum Types) in C#"
     let [<Literal>] AssemblyVersion = "1.0.0"
     let [<Literal>] AssemblyFileVersion = "1.0.0"