+++ src/CSharp.UnionTypes.VSIX/CSharpUnionTypeGenerator.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Runtime.InteropServices;
 using System.Text;
-using BrightSword.CSharpExtensions.DiscriminatedUnion;
+using CSharp.UnionTypes;
 using Microsoft.VisualStudio;
 using Microsoft.VisualStudio.OLE.Interop;
 using Microsoft.VisualStudio.Shell;