+++ src/CSharp.UnionTypes/AssemblyInfo.fs
@@ -2,16 +2,16 @@
 namespace System
 open System.Reflection
 
-[<assembly: AssemblyTitleAttribute("BrightSword.CSharpExtensions.DiscriminatedUnion")>]
-[<assembly: AssemblyProductAttribute("BrightSword.CSharpExtensions.DiscriminatedUnion")>]
+[<assembly: AssemblyTitleAttribute("CSharp.UnionTypes")>]
+[<assembly: AssemblyProductAttribute("CSharp.UnionTypes")>]
 [<assembly: AssemblyDescriptionAttribute("DSL, Parser & Code Generator for Discriminated Union Types (Sum Types) in C#")>]
 [<assembly: AssemblyVersionAttribute("1.0.0")>]
 [<assembly: AssemblyFileVersionAttribute("1.0.0")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] AssemblyTitle = "BrightSword.CSharpExtensions.DiscriminatedUnion"
-    let [<Literal>] AssemblyProduct = "BrightSword.CSharpExtensions.DiscriminatedUnion"
+    let [<Literal>] AssemblyTitle = "CSharp.UnionTypes"
+    let [<Literal>] AssemblyProduct = "CSharp.UnionTypes"
     let [<Literal>] AssemblyDescription = "DSL, Parser & Code Generator for Discriminated Union Types (Sum Types) in C#"
     let [<Literal>] AssemblyVersion = "1.0.0"
     let [<Literal>] AssemblyFileVersion = "1.0.0"